requirements.txt
absl-py==2.3.1
addict==2.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
av==16.0.1
babel==2.17.0
beautifulsoup4==4.14.2
bidict==0.23.1
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
cloudpickle==3.1.1
colorlog==6.10.1
comet_ml==3.54.2
comm==0.2.3
ConfigArgParse==1.7.1
configobj==5.0.9
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dash==3.3.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
dill==0.4.0
docstring_parser==0.17.0
dulwich==0.24.10
embreex==2.17.7.post7
etils==1.13.0
everett==3.1.0
exceptiongroup==1.3.0
executing==2.2.1
Farama-Notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
fonttools==4.60.1
fpsample==1.0.0
fqdn==1.5.1
fsspec==2025.9.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
google-auth==2.41.1
grpcio==1.76.0
gsplat==1.5.3
gym==0.25.1
gym-notebook-wrapper==1.3.3
gym-notices==0.1.0
gymnasium==1.2.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.3.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
lightning-utilities==0.15.2
lxml==6.0.2
manifold3d==3.3.2
mapbox_earcut==2.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
mistune==3.1.4
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
msgpack-numpy==0.4.8
msgspec==0.20.0
mujoco==3.3.7
multiprocess==0.70.18
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nerfacc==0.5.2
-e git+https://github.com/nerfstudio-project/nerfstudio.git@50e0e3c70c775e89333256213363badbf074f29d#egg=nerfstudio
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nodeenv==1.9.1
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.1.2
nuscenes-devkit==1.1.9
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
open3d==0.19.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathos==0.3.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.5.0
plyfile==1.1.3
pox==0.3.6
ppft==1.7.7
proglog==0.1.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycocotools==2.0.10
pycollada==0.9.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymeshlab==2025.7
pyngrok==7.5.0
PyOpenGL==3.1.10
pyparsing==3.2.5
pyquaternion==0.9.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-json-logger==4.0.0
python-socketio==5.15.0
pytorch-msssim==1.0.0
pytz==2025.2
PyVirtualDisplay==3.0
PyYAML==6.0.3
pyzmq==27.1.0
rawpy==0.25.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
rtree==1.4.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.46.0
shapely==2.1.2
shtab==1.8.0
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
splines==0.3.0
stack-data==0.6.3
svg.path==7.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorly==0.9.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.6.7
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.10.0
triton==3.1.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
vhacdx==0.0.9
viser==1.0.0
wadler_lindig==0.1.7
wandb==0.23.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==2.0.1
wsproto==1.3.2
wurlitzer==3.1.1
xatlas==0.0.11
xxhash==3.6.0
yourdfpy==0.0.58
zipp==3.23.0
